Use sync.Pool instead of each time allocation
We allocate 4kb of space each time and this space won't be collected by GC until
we remove references to syscall.NetlinkMessages, because it stores parts
of original slice. Also allocation of 4kb is unneeded every time and in
common cases makes call slower.
This patch though makes receiving big netlink messages slightly slower(because
of copying answer to new slice), but allows to save significant amount of memory.

Signed-off-by: Alexander Morozov <lk4d4@docker.com>