Invalidate NetlinkSocket fd after close
An issue happens when netlink.LinkSubscribe is called back to back, even
when there is a Close() of the netlink socket (via the done chan
parameter). The issue seems to stem from the fact that after the
close/open happen together, the same fd is allocated. This lets the
s.Receive() calls succeed in more than one goroutine, which causes the
updates to issue randomly since the same fd has multiple readers and the
original Subscribe never finishes.

After a Close(), mark the Socket as invalid (-1) so that calls to
Send()/Receive() will fail going forward.

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>