addr: don't log errors happening when receiving notifications
A user may have its own log handling. Another solution would be to
make this configurable by allowing a user to provide a logger. A
better solution would be to transmit the errors to some handler to let
the package user knows about them (notably the nature of the receive
error).

This is not done here since this is the only use of `log.Printf`. For
example, when listening for routes, errors are not displayed.