Service curves fix
There was something off with the math for the ServiceCurves. The set value was 8
times too large and the returned one was 8 times too small.

This is now fixed in `class.go` and `class_linux.go` so the user can just set the rate in bit and
it will match the tc show output.