Service curves fix + expansion
Added a function that simply sets a rate on a curve, so the
behaviour oc tc class add .. sc/ls/ul rate is mimmicked.

There was something of with the math for the ServiceCurves.
The set value was 8 times to large and the returned one was
8 times to small. This is now fixed in class_linux.go so
the user can just set the rate in bit and it will match the
tc show output