rule: Changes IPProto type to u8
netlink defines[1] the type of FRA_IP_PROTO to be a u8 but the definition in
rule.go treats this field as a u32. This causes messages to appear in
kernel logs with "attribute type 22 has an invalid length." This is
exercised in TestRuleAddDel, although it does not cause the test to
fail because netlink parses it permissively.

The type of `IPProto` in `type Rule` is left as int to avoid a breaking
API change. Instead, it is truncated in userspace before sending to the
kernel. Truncation is how the kernel handles larger values, so we don't
expect this to change semantics.

[1]: https://elixir.bootlin.com/linux/v6.19.14/source/net/core/fib_rules.c#L862

Co-authored-by: conjones <connerj@cloudflare.com>