Fix reading of rule type and add l3mdev flag.
When reading rules, makes sure to populate the type field of Rule.
Note that an unspecified type is interpreted as UNICAST, for example,
when adding a rule. To make sure that a rule with unspecified type,
which was just added, compares equal to that rule being read back,
we we update Rule.Equal to treat unspecified and UNICAST types as equal.

Also adds a l3mdev flag to Rule, reads it when listing rules,
sets it when updating rules, and makes sure it is taken into
account when comparing rules using Equal as introduced in #1095.