vdpa: Introduce VDPASetAttr, support setting mac
The Linux kernel supports setting mac addresses to existing vdpa devices
since 6.12. The feature was introduced by commit 2f87e9cf.

This commit brings support for that feature to the netlink library by
introducing the VDPASetAttr function.

Note that even though setting the mac address is the only supported
operation currently, the implementation has been left open for MTU,
MaxVQP and Feature setting. In other words, it relies on whether the
kernel supports it or not, and just feeds the kernel's error back to the
caller. That way, no further changes will be needed in the future, if
support for those operations is implemented. Well, apart from tests, as
they currently assume those are not supported.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>